requirements.txt
fastapi==0.109.0
pymongo==4.6.1
python-dotenv==1.0.0
uvicorn==0.25.0
requests==2.31.0

